frontend-app-api: rename extractRouteInfoFromInstanceTree -> extractRouteInfoFromAppNode
